refactor: folder strukture
