+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import Layout from '@/components/Layout';
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
 import { Provider } from 'react-redux';
-import { store } from './store/store';
+import { store } from '../store/store';
 import '../firebase';
 
 const App = ({ Component, pageProps }: AppProps) => {
